refactor: x와 popply usecase 인터페이스 분리
