+++ src/main/kotlin/com/wheretopop/application/popup/PopplyExternalUseCases.kt
@@ -4,9 +4,7 @@ import com.wheretopop.domain.popup.Popup
 import com.wheretopop.domain.popup.PopupId
 import com.wheretopop.infrastructure.popup.external.popply.PopupDetail
 
-interface PopupUseCase {
-    suspend fun crawlPopplyAndSave()
+interface PopplyUseCase {
     suspend fun crawlPopply(): List<PopupDetail>
     suspend fun savePopply(popupDetail: PopupDetail, popupId: PopupId)
-    suspend fun crawlXAndSave(popup: Popup)
 }
