+++ src/main/kotlin/com/wheretopop/application/popup/PopupFacade.kt
@@ -15,19 +15,12 @@ import org.springframework.stereotype.Service
 @Service
 class PopupFacade(
     private val popupService: PopupService,
-    private val popupUseCase: PopupUseCase,
-    private val buildingService: BuildingService
+    private val buildingService: BuildingService,
+    private val popplyUseCase: PopplyUseCase,
+    private val xUseCase: XUseCase
 ) {
-    /**
-     * 무슨 메서드가 필요할까?
-     */
-
-//    suspend fun ingestPopupExternalData() {
-//        return popupPopplyUseCase.crawlPopplyAndSave()
-//    }
-
     suspend fun ingestPopupExternalData() {
-        val popplyPopupDetails = popupUseCase.crawlPopply()
+        val popplyPopupDetails = popplyUseCase.crawlPopply()
         popplyPopupDetails.forEach{ popupDetail: PopupDetail ->
             if (popupDetail.latitude == null || popupDetail.longitude == null) return
             val buildingCommand = BuildingCommand.CreateBuildingCommand(popupDetail.address, Location(popupDetail.latitude, popupDetail.longitude))
@@ -38,8 +31,8 @@ class PopupFacade(
                 buildingId = buildingId
             )
             popupService.savePopup(newPopup)
-            popupUseCase.savePopply(popupDetail, newPopup.id)
-            popupUseCase.crawlXAndSave(newPopup)
+            popplyUseCase.savePopply(popupDetail, newPopup.id)
+            xUseCase.crawlXAndSave(newPopup)
         }
     }
 }
