+++ src/main/kotlin/com/wheretopop/infrastructure/popup/external/XExternalManager.kt
@@ -1,25 +1,17 @@
 package com.wheretopop.infrastructure.popup.external
 
-import com.wheretopop.application.popup.PopupUseCase
+import com.wheretopop.application.popup.PopplyUseCase
+import com.wheretopop.application.popup.XUseCase
 import com.wheretopop.domain.popup.Popup
 import com.wheretopop.domain.popup.PopupId
 import com.wheretopop.infrastructure.popup.external.popply.PopupDetail
 import org.springframework.stereotype.Component
 
 @Component
-class PopupExternalManager(
+class XExternalManager(
     private val popupExternalStore: PopupExternalStore,
     private val popupExternalReader: PopupExternalReader
-): PopupUseCase {
-    override suspend fun crawlPopplyAndSave() {
-        return popupExternalStore.crawlPopplyAndSave()
-    }
-    override suspend fun crawlPopply():List<PopupDetail> {
-        return popupExternalStore.crawlPopply()
-    }
-    override suspend fun savePopply(popupDetail: PopupDetail, popupId: PopupId) {
-        popupExternalStore.savePopply(popupDetail, popupId)
-    }
+): XUseCase {
     override suspend fun crawlXAndSave(popup: Popup) {
         popupExternalStore.crawlXAndSave(popup)
     }
